core/requirements.txt
django
djangorestframework
markdown       
django-filter
python-decouple
gunicorn
black
psycopg2-binary
Pillow
djangorestframework-simplejwt